Refactor some of the data storage and retrieval to simplify display issues
